🚨 fix clippy warnings
